nectar/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Image Generation
diffusers==0.26.3
transformers==4.37.2
accelerate==0.26.1
torch==2.2.0
torchvision==0.17.0
xformers==0.0.24
safetensors==0.4.2
pillow==10.2.0
opencv-python==4.9.0.80

# ComfyUI API Client
websocket-client==1.7.0
aiohttp==3.9.3
requests==2.31.0

# LLM Integration
openai==1.12.0
anthropic==0.18.1

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
pyyaml==6.0.1
